requirements.txt
attrs==19.1.0
